other(refactor): make project community friendly
* groovy connector has been moved into a dedicated project
* project has been refactored using nez archetype template
* master branche and pull requests are built using github actions

[STUDIO-3660](https://bonitasoft.atlassian.net/browse/STUDIO-3660)
