+++ README.MD
@@ -1,0 +1,26 @@
+# Shell Connector
+
+[![Actions Status](https://github.com/bonitasoft/bonita-connector-shell-script/workflows/Build/badge.svg)](https://github.com/bonitasoft/bonita-connector-shell-script/actions?query=workflow%3ABuild)
+[![Sonarcloud Status](https://sonarcloud.io/api/project_badges/measure?project=bonitasoft_bonita-connector-shell-script&metric=alert_status)](https://sonarcloud.io/dashboard?id=bonitasoft_bonita-connector-shell-script)
+[![GitHub release](https://img.shields.io/github/v/release/bonitasoft/bonita-connector-shell-script?color=blue&label=Release)](https://github.com/bonitasoft/bonita-connector-shell-script/releases)
+[![Maven Central](https://img.shields.io/maven-central/v/org.bonitasoft.connectors/bonita-connector-shell-script.svg?label=Maven%20Central&color=orange)](https://search.maven.org/search?q=g:%22org.bonitasoft.connectors%22%20AND%20a:%22bonita-connector-shell-script%22)
+[![License: GPL v2](https://img.shields.io/badge/License-GPL%20v2-yellow.svg)](https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html)
+
+## Build
+
+__Clone__ or __fork__ this repository, then at the root of the project run:
+
+`./mvnw`
+
+## Release
+
+In order to create a new release push a `release-<version>` branch with the desired version in pom.xml.
+Update the `master` with the next SNAPSHOT version.
+
+## Contributing
+
+We would love you to contribute, pull requests are welcome! Please see the [CONTRIBUTING.md](CONTRIBUTING.md) for more information.
+
+## License
+
+The sources and documentation in this project are released under the [GPLv2 License](LICENSE)
