--- groovyScript-assembly.xml
@@ -1,41 +1,0 @@
-<assembly
-    xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-    <id>groovyScript</id>
-    <formats>
-        <format>zip</format>
-    </formats>
-    <includeBaseDirectory>false</includeBaseDirectory>
-    <fileSets>
-        <fileSet>
-            <outputDirectory>/</outputDirectory>
-            <directory>target/classes</directory>
-            <includes>
-                <include>scripting-groovy-script.impl</include>
-                <include>/classpath</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <outputDirectory>/classpath</outputDirectory>
-            <directory>${basedir}/target</directory>
-            <includes>
-                <include>*-sources.jar</include>
-            </includes>
-        </fileSet>
-        <fileSet>
-            <outputDirectory>/classpath</outputDirectory>
-            <directory>${basedir}/target</directory>
-            <includes>
-                <include>*:jar</include>
-            </includes>
-        </fileSet>
-    </fileSets>
-    <dependencySets>
-        <dependencySet>
-            <unpack>false</unpack>
-            <scope>runtime</scope>
-            <outputDirectory>/classpath</outputDirectory>
-        </dependencySet>
-    </dependencySets>
-</assembly>
