+++ package.json
@@ -1,0 +1,4 @@
+{
+  "name": "bonita-connector-shell",
+  "version": "1.1.1"
+}
