+++ src/assembly/all-assembly.xml
@@ -1,0 +1,34 @@
+<assembly
+        xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
+    <id>all</id>
+    <formats>
+        <format>zip</format>
+    </formats>
+    <includeBaseDirectory>false</includeBaseDirectory>
+    <fileSets>
+        <fileSet>
+            <outputDirectory/>
+            <directory>target/classes</directory>
+            <includes>
+                <!-- Only first level files -->
+                <include>*.*</include>
+            </includes>
+        </fileSet>
+        <fileSet>
+            <outputDirectory>classpath</outputDirectory>
+            <directory>${basedir}/target</directory>
+            <includes>
+                <include>*:jar</include>
+            </includes>
+        </fileSet>
+    </fileSets>
+    <dependencySets>
+        <dependencySet>
+            <unpack>false</unpack>
+            <scope>runtime</scope>
+            <outputDirectory>classpath</outputDirectory>
+        </dependencySet>
+    </dependencySets>
+</assembly>
