+++ src/assembly/shell-assembly.xml
