--- src/main/java/org/bonitasoft/connectors/scripting/GroovyConnector.java
@@ -1,40 +1,0 @@
-/**
- * Copyright (C) 2009 BonitaSoft S.A.
- * BonitaSoft, 31 rue Gustave Eiffel - 38000 Grenoble
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2.0 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.bonitasoft.connectors.scripting;
-
-import org.bonitasoft.engine.connector.AbstractConnector;
-import org.bonitasoft.engine.connector.ConnectorException;
-import org.bonitasoft.engine.connector.ConnectorValidationException;
-
-/**
- * @author Baptiste Mesta
- * @author Frédéric Bouquet
- */
-public class GroovyConnector extends AbstractConnector {
-
-    private static final String SCRIPT = "script";
-
-    private static final String RESULT = "result";
-
-    @Override
-    protected void executeBusinessLogic() throws ConnectorException {
-        setOutputParameter(RESULT, getInputParameter(SCRIPT));
-    }
-
-    @Override
-    public void validateInputParameters() throws ConnectorValidationException {
-
-    }
-}
