+++ src/main/java/org/bonitasoft/connectors/shell/ShellConnector.java
@@ -11,7 +11,7 @@
  * program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth
  * Floor, Boston, MA 02110-1301, USA.
  **/
-package org.bonitasoft.connectors.scripting;
+package org.bonitasoft.connectors.shell;
 
 import static java.util.Optional.ofNullable;
