+++ src/main/resources-filtered/scripting-shell.def
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <definition:ConnectorDefinition xmlns:definition="http://www.bonitasoft.org/ns/connector/definition/6.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<id>scripting-shell</id>
-	<version>1.0.0</version>
+	<id>${shell.def.id}</id>
+	<version>${shell.def.version}</version>
 	<icon>shell.png</icon>
-
 	<category id="Script" icon="scripting.png" />
 	
 	<input name="interpreter" type="java.lang.String" mandatory="true" />
