--- src/main/resources/scripting-groovy-script.def
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<definition:ConnectorDefinition xmlns:definition="http://www.bonitasoft.org/ns/connector/definition/6.1"
-                                xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<id>scripting-groovy-script</id>
-	<version>1.0.1</version>
-	<icon>groovy.png</icon>
-
-	<category id="Script" icon="scripting.png" />
-
-	<input name="script" type="java.lang.String" mandatory="true" />
-	<input name="variables" type="java.util.List" />
-
-	<output name="result" type="java.lang.Object" />
-
-	<page id="groovy">
-		<widget id="script" inputName="script" xsi:type="definition:ScriptEditor" interpreter="GROOVY"/>
-	</page>
-</definition:ConnectorDefinition>
