--- src/main/resources/scripting-groovy-script.impl
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<implementation:connectorImplementation xmlns:implementation="http://www.bonitasoft.org/ns/connector/implementation/6.0">
-
-	<definitionId>scripting-groovy-script</definitionId>
-	<definitionVersion>1.0.1</definitionVersion>
-    <description>Support Groovy 2.2.4</description>
-	<implementationClassname>org.bonitasoft.connectors.scripting.GroovyScriptConnector</implementationClassname>
-	<implementationId>scripting-groovy-script-impl</implementationId>
-	<implementationVersion>$implementation.version$</implementationVersion>
-
-$Dependencies$
-
-</implementation:connectorImplementation>
