--- src/main/resources/scripting-groovy-script_de.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Skript
-connectorDefinitionLabel=Groovy 2.4
-connectorDefinitionDescription=Groovy-Skript 2.4 ausf\u00fchren
-
-groovy.pageTitle=Groovy-Skript
-groovy.pageDescription=Skript in das Feld unten eingeben
-
-script.label=Skript
-script.description=
