--- src/main/resources/scripting-groovy-script_en.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4
-connectorDefinitionDescription=Execute a Groovy 2.4 script outside the Bonita transaction. There is no time limit on execution.
-
-groovy.pageTitle=Groovy script
-groovy.pageDescription=Enter the script in the field below
-
-script.label=Script
-script.description=
