--- src/main/resources/scripting-groovy-script_es.properties
@@ -1,11 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4
-connectorDefinitionDescription=Ejecutar un script Groovy 2.4 fuera de una transacci\u00f3n de Bonita. No hay l\u00edmite de tiempo para la ejecuci\u00f3n. 
-
-
-groovy.pageTitle=Groovy script
-groovy.pageDescription=Introduzca la secuencia de comandos en el campo de abajo
-
-script.label=Script
-script.description=
