--- src/main/resources/scripting-groovy-script_fr.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4
-connectorDefinitionDescription=Ex\u00e9cute un script Groovy 2.4 en dehors d'un transaction Bonita. (pas de limite de temps lors de l'ex\u00e9cution)
-
-groovy.pageTitle=Script Groovy
-groovy.pageDescription=Entrer ici le script \u00e0 ex\u00e9cuter
-
-script.label=Script
-script.description=
