--- src/main/resources/scripting-groovy-script_it.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4 
-connectorDefinitionDescription=Esegui uno script Groovy 2.4
-
-groovy.pageTitle=script groovy 
-groovy.pageDescription=Inserisci lo script nel campo sottostante
-
-script.label=Script
-script.description=
