--- src/main/resources/scripting-groovy.impl
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<implementation:connectorImplementation xmlns:implementation="http://www.bonitasoft.org/ns/connector/implementation/6.0">
-
-	<definitionId>scripting-groovy</definitionId>
-	<definitionVersion>1.0.2</definitionVersion>
-    <description>Support Groovy 2.4.4</description>
-	<implementationClassname>org.bonitasoft.connectors.scripting.GroovyConnector</implementationClassname>
-	<implementationId>scripting-groovy-impl</implementationId>
-	<implementationVersion>$implementation.version$</implementationVersion>
-
-$Dependencies$
-
-</implementation:connectorImplementation>
