--- src/main/resources/scripting-groovy_en.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4 (deprecated)
-connectorDefinitionDescription=Execute a Groovy 2.4 script.\nThe connector executes the script within the Bonita transaction. If script execution takes longer than the Bonita transaction time, the script fails.
-
-groovy.pageTitle=Groovy script
-groovy.pageDescription=Enter the script in the field below
-
-script.label=Script
-script.description=
