--- src/main/resources/scripting-groovy_es.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4 (deprecated)
-connectorDefinitionDescription=Ejecutar un script Groovy 2.4\n El conector ejecuta el script dentro de una transacci\u00f3n de Bonita, Si la ejecuci\u00f3n del script supera el tiempo de una transacci\u00f3n de Bonita, el script fallar\u00e1.
-
-groovy.pageTitle=Groovy script
-groovy.pageDescription=Introduzca la secuencia de comandos en el campo de abajo
-
-script.label=Script
-script.description=
