--- src/main/resources/scripting-groovy_fr.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=Groovy 2.4 (d\u00e9pr\u00e9ci\u00e9)
-connectorDefinitionDescription=Ex\u00e9cute un script Groovy 2.4 au sein d'une transaction Bonita.\n Attention, si le temps d'ex\u00e9cution du script d\u00e9passe celui de la transaction, le script \u00e9choue.
-
-groovy.pageTitle=Script Groovy
-groovy.pageDescription=Entrer ici le script \u00e0 ex\u00e9cuter
-
-script.label=Script
-script.description=
