--- src/main/resources/scripting-groovy_ja.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=\u30b9\u30af\u30ea\u30d7\u30c8
-connectorDefinitionLabel=Groovy 2.4 (deprecated)
-connectorDefinitionDescription=Groovy 2.4  \u306e\u30b9\u30af\u30ea\u30d7\u30c8\u3092\u5b9f\u884c\u3057\u307e\u3059
-
-groovy.pageTitle=Groovy\u30b9\u30af\u30ea\u30d7\u30c8
-groovy.pageDescription=\u4ee5\u4e0b\u306e\u30d5\u30a3\u30fc\u30eb\u30c9\u306b\u30b9\u30af\u30ea\u30d7\u30c8\u3092\u5165\u529b\u3057\u3066\u304f\u3060\u3055\u3044\u3002
-
-script.label=\u30b9\u30af\u30ea\u30d7\u30c8
-script.description=[\u7a7a\u306e\u6587\u5b57\u5217]
