--- src/main/resources/scripting-groovy_pt.properties
@@ -1,10 +1,0 @@
-#X-Generator: crowdin.net
-Script.category=Script
-connectorDefinitionLabel=\ Groovy 2.4 (deprecated)
-connectorDefinitionDescription=Executa um script Groovy 2.4
-
-groovy.pageTitle=Script Groovy
-groovy.pageDescription=Digite o script no campo abaixo
-
-script.label=Script
-script.description=
