--- src/main/resources/scripting-shell.impl
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<implementation:connectorImplementation xmlns:implementation="http://www.bonitasoft.org/ns/connector/implementation/6.0">
-
-	<definitionId>scripting-shell</definitionId>
-	<definitionVersion>1.0.0</definitionVersion>
-	<description>Support scripts or commands on any kind of system</description>
-	<implementationClassname>org.bonitasoft.connectors.scripting.ShellConnector</implementationClassname>
-	<implementationId>scripting-shell-impl</implementationId>
-	<implementationVersion>$implementation.version$</implementationVersion>
-
-$Dependencies$
-
-</implementation:connectorImplementation>
