+++ src/test/java/org/bonitasoft/connectors/shell/ShellConnectorTest.java
@@ -11,7 +11,7 @@
  * program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth
  * Floor, Boston, MA 02110-1301, USA.
  **/
-package org.bonitasoft.connectors.scripting;
+package org.bonitasoft.connectors.shell;
 
 import static java.lang.String.join;
 import static java.lang.System.lineSeparator;
@@ -23,6 +23,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Logger;
 
+import org.bonitasoft.connectors.shell.ShellConnector;
 import org.bonitasoft.engine.connector.ConnectorException;
 import org.bonitasoft.engine.connector.ConnectorValidationException;
 import org.junit.Test;
